Split composite examples
